requirements.txt
# ── Core ───────────────────────────────────────────────────────────────────────
opencv-python>=4.8.0
numpy>=1.24.0

# ── Detection ──────────────────────────────────────────────────────────────────
# Option A — YOLOv8 (recommended, easier install):
ultralytics>=8.0.0
# Option B — YOLOv5 via torch hub (comment out ultralytics above and uncomment):
# torch>=2.0.0
# torchvision>=0.15.0

# ── Tracking ───────────────────────────────────────────────────────────────────
deep-sort-realtime>=1.3.2

# ── OCR ────────────────────────────────────────────────────────────────────────
easyocr>=1.7.0

# ── PDF Challan ────────────────────────────────────────────────────────────────
fpdf2>=2.7.0

# ── Dashboard ──────────────────────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0

# ── Utilities ──────────────────────────────────────────────────────────────────
pillow>=10.0.0
scipy>=1.11.0
